requirements.txt
tweepy>=4.14.0
dash>=2.14.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow-cpu>=2.15.0
joblib>=1.3.0
python-dotenv>=1.0.0
orjson>=3.9.0
gunicorn>=21.2.0
pytest>=7.4.0
